refac/ 통계 기능 앱 추가 및 코드 이동
