+++ Wayne_Hills/statistic/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class StatisticConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'statistic'
