+++ Wayne_Hills/statistic/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from statistic import views
+
+# statistic/
+urlpatterns = [
+    path('gender', views.GenderStatisticsView.as_view()),
+]
